Bug fix for adding new management actions.
What was previously stored in an object called library_config was moved to an object called "store". In this new store object, currentLibrary.id didn't exist, so when the user tried to add a new management action, a javascript error was thrown. All references to store[currentLibrary.id] were changed to store[currentLibrary.name].